admin tasks, orders

+++ pkg/service/option.go
@@ -17,12 +17,16 @@ func (s *OptionsService) Create(parentId int, option taskexchange.Option) (int, 
 	return s.repo.Create(parentId, option)
 }
 
-func (s *OptionsService) GetAll() ([]taskexchange.Option, error) {
-	return s.repo.GetAll()
+func (s *OptionsService) GetAll(full bool) ([]taskexchange.Option, error) {
+	return s.repo.GetAll(full)
 }
 
-func (s *OptionsService) GetById(id int) (taskexchange.Option, error) {
-	return s.repo.GetById(id)
+func (s *OptionsService) GetCategories() ([]taskexchange.Option, error) {
+	return s.repo.GetCategories()
+}
+
+func (s *OptionsService) GetById(id int, full bool) (taskexchange.Option, error) {
+	return s.repo.GetById(id, full)
 }
 
 func (s *OptionsService) Update(id int, input taskexchange.UpdateOptionInput) error {
@@ -34,5 +38,14 @@ func (s *OptionsService) Update(id int, input taskexchange.UpdateOptionInput) er
 }
 
 func (s *OptionsService) Delete(id int) error {
+	option, err := s.repo.GetById(id, true)
+	if err != nil {
+		return err
+	}
+
+	if option.DeletedAt != nil {
+		return s.repo.Restore(id)
+	}
+
 	return s.repo.Delete(id)
 }
